refactor: Move skeletons to new ui/views directory
