+++ apps/web/src/app/dangle/loading.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { DangleRecommendSkeleton } from "@/components/Skeletons/DangleSkeletons";
+import { DangleRecommendSkeleton } from "@/ui/views/Skeletons/DangleSkeletons";
 
 export default function Loading() {
   return <DangleRecommendSkeleton />;
