+++ apps/web/src/app/detail/[contentId]/loading.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { DetailSkeleton } from "@/components/Skeletons/DetailSkeletons";
+import { DetailSkeleton } from "@/ui/views/Skeletons/DetailSkeletons";
 
 export default function Loading() {
   return <DetailSkeleton />;
