+++ apps/web/src/app/list/loading.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { ListSkeleton } from "@/components/Skeletons/ListSkeleton";
+import { ListSkeleton } from "@/ui/views/Skeletons/ListSkeleton";
 
 export default function ListLoadingPage() {
   return <ListSkeleton />;
