+++ apps/web/src/app/search/loading.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { SearchPageSkeleton } from "@/components/Skeletons/SearchPageSkeleton";
+import { SearchPageSkeleton } from "@/ui/views/Skeletons/SearchPageSkeleton";
 
 export default function Loading() {
   return <SearchPageSkeleton />;
