+++ apps/web/src/ui/views/Skeletons/AiCareSkeleton.tsx
