+++ apps/web/src/ui/views/Skeletons/DetailSkeletons.tsx
