+++ apps/web/src/ui/views/Skeletons/SearchPageSkeleton.tsx
